+++ python/GafferUI/NumericPlugValueWidget.py
@@ -179,13 +179,7 @@ class NumericPlugValueWidget( GafferUI.PlugValueWidget ) :
 					if Gaffer.Animation.isAnimated( plug ) :
 						curve = Gaffer.Animation.acquire( plug )
 						if self.__numericWidget.getText() != self.__numericWidget.valueToString( curve.evaluate( self.getContext().getTime() ) ) :
-							curve.addKey(
-								Gaffer.Animation.Key(
-									self.getContext().getTime(),
-									self.__numericWidget.getValue(),
-									Gaffer.Animation.Interpolation.Linear
-								)
-							)
+							curve.insertKey( self.getContext().getTime(), self.__numericWidget.getValue() )
 					else :
 						try :
 							plug.setValue( self.__numericWidget.getValue() )