+++ src/GafferSceneUI/TransformTool.cpp
@@ -1251,8 +1251,8 @@ bool TransformTool::keyPress( const GafferUI::KeyEvent &event )
 					if( Animation::canAnimate( plug.get() ) )
 					{
 						const float value = plug->getValue();
-						Animation::CurvePlug *curve = Animation::acquire( plug.get() );
-						curve->addKey( new Animation::Key( s.context()->getTime(), value ) );
+						Animation::CurvePlug* const curve = Animation::acquire( plug.get() );
+						curve->insertKey( s.context()->getTime(), value );
 					}
 				}
 			}
@@ -1286,7 +1286,7 @@ void TransformTool::setValueOrAddKey( Gaffer::FloatPlug *plug, float time, float
 	if( Animation::isAnimated( plug ) )
 	{
 		Animation::CurvePlug *curve = Animation::acquire( plug );
-		curve->addKey( new Animation::Key( time, value ) );
+		curve->insertKey( time, value );
 	}
 	else
 	{